+++ app/package.json
@@ -10,7 +10,9 @@
   },
   "dependencies": {
     "@expo/webpack-config": "^18.0.1",
-    "expo-app-loading": "^2.1.1",
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
+    "expo-app-loading": "^1.3.0",
     "expo-font": "~11.1.1",
     "expo-linear-gradient": "~12.1.2",
     "expo-splash-screen": "~0.18.2",
@@ -18,7 +20,11 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.71.7",
-    "react-native-web": "~0.18.10"
+    "react-native-navigation": "^7.33.0-alpha.8",
+    "react-native-safe-area-context": "^4.5.0",
+    "react-native-screens": "~3.20.0",
+    "react-native-web": "~0.18.10",
+    "react-navigation": "^4.4.4"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
